作品情報の更新
登録した作品画像の削除処理の作成
