+++ src/main/java/com/jp/wonfes/work/controller/form/WorkEditForm.java
@@ -1,48 +1,48 @@
 package com.jp.wonfes.work.controller.form;
 
 public class WorkEditForm extends WorkRegistForm{
-	private Integer productId;
-	private String workImg1DelFlg;
-	private String workImg2DelFlg;
-	private String workImg3DelFlg;
-	private String workImg4DelFlg;
-	private String workImg5DelFlg;
-	
-	public Integer getProductId() {
-		return productId;
-	}
-	public void setProductId(Integer productId) {
-		this.productId = productId;
-	}
-	public String getWorkImg1DelFlg() {
-		return workImg1DelFlg;
-	}
-	public void setWorkImg1DelFlg(String workImg1DelFlg) {
-		this.workImg1DelFlg = workImg1DelFlg;
-	}
-	public String getWorkImg2DelFlg() {
-		return workImg2DelFlg;
-	}
-	public void setWorkImg2DelFlg(String workImg2DelFlg) {
-		this.workImg2DelFlg = workImg2DelFlg;
-	}
-	public String getWorkImg3DelFlg() {
-		return workImg3DelFlg;
-	}
-	public void setWorkImg3DelFlg(String workImg3DelFlg) {
-		this.workImg3DelFlg = workImg3DelFlg;
-	}
-	public String getWorkImg4DelFlg() {
-		return workImg4DelFlg;
-	}
-	public void setWorkImg4DelFlg(String workImg4DelFlg) {
-		this.workImg4DelFlg = workImg4DelFlg;
-	}
-	public String getWorkImg5DelFlg() {
-		return workImg5DelFlg;
-	}
-	public void setWorkImg5DelFlg(String workImg5DelFlg) {
-		this.workImg5DelFlg = workImg5DelFlg;
-	}
+//	private Integer productId;
+//	private String workImg1DelFlg;
+//	private String workImg2DelFlg;
+//	private String workImg3DelFlg;
+//	private String workImg4DelFlg;
+//	private String workImg5DelFlg;
+//	
+//	public Integer getProductId() {
+//		return productId;
+//	}
+//	public void setProductId(Integer productId) {
+//		this.productId = productId;
+//	}
+//	public String getWorkImg1DelFlg() {
+//		return workImg1DelFlg;
+//	}
+//	public void setWorkImg1DelFlg(String workImg1DelFlg) {
+//		this.workImg1DelFlg = workImg1DelFlg;
+//	}
+//	public String getWorkImg2DelFlg() {
+//		return workImg2DelFlg;
+//	}
+//	public void setWorkImg2DelFlg(String workImg2DelFlg) {
+//		this.workImg2DelFlg = workImg2DelFlg;
+//	}
+//	public String getWorkImg3DelFlg() {
+//		return workImg3DelFlg;
+//	}
+//	public void setWorkImg3DelFlg(String workImg3DelFlg) {
+//		this.workImg3DelFlg = workImg3DelFlg;
+//	}
+//	public String getWorkImg4DelFlg() {
+//		return workImg4DelFlg;
+//	}
+//	public void setWorkImg4DelFlg(String workImg4DelFlg) {
+//		this.workImg4DelFlg = workImg4DelFlg;
+//	}
+//	public String getWorkImg5DelFlg() {
+//		return workImg5DelFlg;
+//	}
+//	public void setWorkImg5DelFlg(String workImg5DelFlg) {
+//		this.workImg5DelFlg = workImg5DelFlg;
+//	}
 	
 }
