+++ src/main/java/com/jp/wonfes/work/controller/form/WorkRegistForm.java
@@ -35,6 +35,12 @@ public class WorkRegistForm {
 	private String workImg3Url;
 	private String workImg4Url;
 	private String workImg5Url;
+	private Integer productId;
+	private String workImg1DelFlg;
+	private String workImg2DelFlg;
+	private String workImg3DelFlg;
+	private String workImg4DelFlg;
+	private String workImg5DelFlg;
 	
 	public Integer getDealerId() {
 		return dealerId;
@@ -137,6 +143,42 @@ public class WorkRegistForm {
 	}
 	public void setWorkImg5Url(String workImg5Url) {
 		this.workImg5Url = workImg5Url;
+	}
+	public Integer getProductId() {
+		return productId;
+	}
+	public void setProductId(Integer productId) {
+		this.productId = productId;
+	}
+	public String getWorkImg1DelFlg() {
+		return workImg1DelFlg;
+	}
+	public void setWorkImg1DelFlg(String workImg1DelFlg) {
+		this.workImg1DelFlg = workImg1DelFlg;
+	}
+	public String getWorkImg2DelFlg() {
+		return workImg2DelFlg;
+	}
+	public void setWorkImg2DelFlg(String workImg2DelFlg) {
+		this.workImg2DelFlg = workImg2DelFlg;
+	}
+	public String getWorkImg3DelFlg() {
+		return workImg3DelFlg;
+	}
+	public void setWorkImg3DelFlg(String workImg3DelFlg) {
+		this.workImg3DelFlg = workImg3DelFlg;
+	}
+	public String getWorkImg4DelFlg() {
+		return workImg4DelFlg;
+	}
+	public void setWorkImg4DelFlg(String workImg4DelFlg) {
+		this.workImg4DelFlg = workImg4DelFlg;
+	}
+	public String getWorkImg5DelFlg() {
+		return workImg5DelFlg;
+	}
+	public void setWorkImg5DelFlg(String workImg5DelFlg) {
+		this.workImg5DelFlg = workImg5DelFlg;
 	}
